[Silver V] Title: 공포의 면담실, Time: 568 ms, Memory: 86000 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/30088. 공포의 면담실/공포의 면담실.java
@@ -8,7 +8,7 @@ class Main{
 	public static void main(String[] args)throws Exception{
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		int N		= Integer.parseInt(br.readLine());
-		int psum[]	= new int[N];
+		long psum[]	= new long[N];
 		
 		for(int i=0; i<N; i++)
 		{
@@ -20,7 +20,7 @@ class Main{
 		
 		Arrays.sort(psum);
 		
-		int sum = psum[0];
+		long sum = psum[0];
 		for(int i=1; i<N; i++)
 			sum += psum[i] += psum[i-1];
